pizerow/requirements.txt
paho-mqtt>=2.0,<3.0
smbus2>=0.4,<1.0
